src/agent-core/docker-graph/requirements.txt
langgraph==1.0.10
strands-agents==1.21.0
bedrock-agentcore==1.1.4
mcp-proxy-for-aws==1.1.6
pydantic==2.12.3
boto3==1.42.26
aws-opentelemetry-distro==0.14.1
